backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
torch>=2.0.0
torchvision>=0.15.0
python-docx>=1.1.0
pydantic>=2.7.4
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.6.1
numpy>=1.26.4
tiktoken>=0.5.2 